--- a/src/Kestrel.Core/Internal/Http/Http1Connection.cs
+++ b/src/Kestrel.Core/Internal/Http/Http1Connection.cs
@@ -405,7 +405,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 
         protected override void OnReset()
         {
-            FastFeatureSet(typeof(IHttpUpgradeFeature), this);
+            ResetIHttpUpgradeFeature();
 
             _requestTimedOut = false;
             _requestTargetForm = HttpRequestTarget.Unknown;