… src/Kestrel.Core/Internal/Http/HttpProtocol.Generated.cs