Speed up the generic `IFeatureCollection` `Get<T>`/`Set<T>`

Pre
```
                 Method |     Mean |         Op/s |
----------------------- |---------:|-------------:|
     GetViaTypeOf_First | 12.96 ns | 77,136,481.9 | 
    GetViaGeneric_First | 22.38 ns | 44,679,211.4 | IHttpRequestFeature
    
      GetViaTypeOf_Last | 69.91 ns | 14,304,944.1 |
     GetViaGeneric_Last | 80.70 ns | 12,392,225.5 | IHttpSendFileFeature
     
    GetViaTypeOf_Custom | 76.84 ns | 13,014,603.2 |
   GetViaGeneric_Custom | 84.15 ns | 11,882,891.0 |
   
  GetViaTypeOf_NotFound | 76.29 ns | 13,108,494.2 |
 GetViaGeneric_NotFound | 84.44 ns | 11,843,124.1 |
```

Post
```
                 Method |     Mean |         Op/s |
----------------------- |---------:|-------------:|
     GetViaTypeOf_First | 12.70 ns | 78,733,231.9 |
    GetViaGeneric_First | 19.96 ns | 50,103,862.6 | IHttpRequestFeature
    
      GetViaTypeOf_Last | 66.28 ns | 15,087,099.9 |
     GetViaGeneric_Last | 33.59 ns | 29,771,919.4 | IHttpSendFileFeature
     
    GetViaTypeOf_Custom | 75.68 ns | 13,214,141.0 |
   GetViaGeneric_Custom | 44.21 ns | 22,617,022.7 |
   
  GetViaTypeOf_NotFound | 75.45 ns | 13,254,289.3 |
 GetViaGeneric_NotFound | 44.53 ns | 22,454,601.1 |
```