@benaadams 2017-08-22
Shame about the `Interface<Generic>` cost, because the generic function itself is pretty good