@benaadams 2017-09-26
Using two arrays of struct wrappers for `Type` and `object` with a for loop (for high density on Type matching and to switch covariance off on the array) rather than a branching if/else is slower than the `typeof` version, so way slower than the generic version for `_Last`.