@davidfowl 2018-02-03
This PR needs to do over. We refactored everything for http2. This change is interesting though, we should look at it again.