@benaadams 2018-02-03
Rebased https://github.com/aspnet/KestrelHttpServer/pull/2290